Fix multiple operator glitch
Made sure the operator button does not evaluate the
equation if no second number has been provided.

This is because if an operator (such as *) was clicked
multiples times in a row, the calculator would set
number2 as an empty string and then evaluate the equation,
returning a faulty result which is no longer accurate
for the equation so far.